fix(scripts): share one healthcheck wait so the drizzle lane inherits the podman timer workaround
The drizzle lane had its own copy of waitHealthy without e2e.mjs's
`podman healthcheck run` fallback, so on a runner where systemd is not
init the reported status sat at 'starting' and all three dialects failed
after 188s with the registry long since ready.
